backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.3
Pillow>=10.0.0
reportlab==4.0.7
openpyxl==3.1.2
python-decouple==3.8
celery==5.3.4
channels==4.0.0
channels-redis==4.1.0
websockets==12.0
redis==5.0.1
django-extensions==3.2.3
psutil==7.0.0

# Dépendances Redis et WebSockets
redis>=5.0.0
channels-redis>=4.1.0
django-redis>=5.3.0
celery>=5.3.0
channels>=4.0.0
websockets>=11.0.0

# Monitoring et performance
psutil>=5.9.0
django-extensions>=3.2.0

# CORS pour les WebSockets
django-cors-headers>=4.3.0
